#2 feat: base build done - ready for debug&test
