+++ scripts/Run_Client.sh
@@ -1,4 +1,4 @@
 # set environment variable - only valid for current terminal session!
 export COMMONAPI_CONFIG="$PWD"/build/etc 
-cd build
+cd ../build
 ./CarInfoClient
