+++ src-gen/core/v0/commonapi/CarInfo.hpp
