+++ src-gen/core/v0/commonapi/CarInfoProxy.hpp
