+++ src-gen/core/v0/commonapi/CarInfoProxyBase.hpp
