+++ src-gen/core/v0/commonapi/CarInfoStub.hpp
