+++ src-gen/core/v0/commonapi/CarInfoStubDefault.hpp
