+++ src-gen/someip/v0/commonapi/CarInfoSomeIPStubAdapter.cpp
