+++ src/CarInfoStubImpl.cpp
@@ -1,5 +1,6 @@
 #include "CarInfoStubImpl.hpp"
 
+
 CarInfoStubImpl::CarInfoStubImpl() {
 }
 
@@ -7,7 +8,7 @@ CarInfoStubImpl::~CarInfoStubImpl() {
 }
 
 void CarInfoStubImpl::setBatteryStruct(float level_, float voltage_, float current_, float consumption_) {
-    CommonTypes::batteryStruct battery;
+    v0::commonapi::CommonTypes::batteryStruct battery;
     battery.setLevel(level_);
     battery.setVoltage(voltage_);
     battery.setCurrent(current_);
